+++ ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/management/viewmodel/ApplicationAssessmentProgressRowViewModel.java
@@ -1,15 +1,15 @@
 package org.innovateuk.ifs.management.viewmodel;
 
 /**
- * TODO
+ * Holder of model attributes for an assessor in the Application Progress view.
  */
 abstract class ApplicationAssessmentProgressRowViewModel {
 
     private String name;
-    private int totalApplicationsCount;
-    private int assignedCount;
+    private long totalApplicationsCount;
+    private long assignedCount;
 
-    protected ApplicationAssessmentProgressRowViewModel(String name, int totalApplicationsCount, int assignedCount) {
+    protected ApplicationAssessmentProgressRowViewModel(String name, long totalApplicationsCount, long assignedCount) {
         this.name = name;
         this.totalApplicationsCount = totalApplicationsCount;
         this.assignedCount = assignedCount;
@@ -19,11 +19,11 @@ abstract class ApplicationAssessmentProgressRowViewModel {
         return name;
     }
 
-    public int getTotalApplicationsCount() {
+    public long getTotalApplicationsCount() {
         return totalApplicationsCount;
     }
 
-    public int getAssignedCount() {
+    public long getAssignedCount() {
         return assignedCount;
     }
 }
